Made a ViewController
ViewController now has an actual view controller. There's an array of
views, you fill those with all possible views. You can then change the
view by calling to the ChangeView function, parameter is the index of a
certain view in the array. In the View class, not much has changed,
since it doesn't contain anything yet.